+++ src/test/java/org/mmarini/railways2/model/StationStatus2SectionsTest.java
@@ -47,6 +47,7 @@ import static org.hamcrest.Matchers.*;
 import static org.junit.jupiter.api.Assertions.*;
 import static org.mmarini.railways2.model.Matchers.isEdge;
 import static org.mmarini.railways2.model.Matchers.isRoute;
+import static org.mmarini.railways2.model.RailwayConstants.COACH_LENGTH;
 import static org.mmarini.railways2.model.RailwayConstants.ENTRY_TIMEOUT;
 import static org.mockito.ArgumentMatchers.anyInt;
 import static org.mockito.Mockito.*;
@@ -101,11 +102,36 @@ class StationStatus2SectionsTest extends WithStationStatusTest {
     }
 
     @Test
+    void createTrainByExitExiting() {
+        // Give ...
+        status = withTrain()
+                .addTrain(new WithTrain.TrainBuilder("ex", 3, "a", "b")
+                        .exiting("b", 3 * COACH_LENGTH - 1))
+                .addTrain(3, "a", "b", "cd", "d", 0)
+                .build();
+
+        // When ...
+        Map<Section, Train> map = status.createTrainBySection();
+
+        // Than ...
+        assertNotNull(map);
+        assertEquals(2, map.size());
+        assertThat(map, hasEntry(
+                hasProperty("id", equalTo("ab")),
+                equalTo(train("ex"))));
+        assertThat(map, hasEntry(
+                hasProperty("id", equalTo("cd")),
+                equalTo(train("TT1"))));
+    }
+
+    @Test
     void createTrainBySection() {
         // Give ...
         status = withTrain()
                 .addTrain(3, "a", "b", "ab", "b", 0)
                 .addTrain(3, "a", "b", "cd", "d", 0)
+                .addTrain(new WithTrain.TrainBuilder("ex", 3, "a", "b")
+                        .exiting("b", 3 * COACH_LENGTH + 1))
                 .build();
 
         // When ...